p4i integration Integrate 259545 from 4.3 to main: Fixes:    Fix Qtopia compilation RevBy:    Jason Details:  Ambiguous function call to qFuzzyCompare.
[git-p4: depot-paths = "//depot/qt/main/": change = 259546]

+++ src/gui/painting/qcolor.cpp
@@ -1368,7 +1368,7 @@ QColor QColor::toHsv() const
     const qreal min = Q_MIN_3(r, g, b);
     const qreal delta = max - min;
     color.ct.ahsv.value = qRound(max * USHRT_MAX);
-    if (qFuzzyCompare(delta, 0.0)) {
+    if (qFuzzyCompare(delta, qreal(0.0))) {
         // achromatic case, hue is undefined
         color.ct.ahsv.hue = USHRT_MAX;
         color.ct.ahsv.saturation = 0;
